Refactor History Ingestor (#24)
- Replace `BTCHistorySyncService` with `BTCHistoryIngesterService` for improved batching and parallelism.
- Introduce the `batcher` package for efficient batch processing.
- Update schemas: `block_height` now `UInt64`, removed `fee` from transactions.
- Add `RandomMissingBlockHeights` to repository interface.
- Rename CLI command and environment configs for history ingestion.
- Update tests to reflect service and schema changes.
- Adjust dependencies and migrate database according to schema changes.

+++ migrations/clickhouse/1_blocks.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS btc_blocks (
     network LowCardinality(String) CODEC(ZSTD(1)),
-    height UInt32 CODEC(ZSTD(1)),
+    height UInt64 CODEC(ZSTD(1)),
     hash FixedString(64) CODEC(ZSTD(1)),
     timestamp DateTime('UTC') CODEC(Delta(4), LZ4),
     version UInt32  CODEC(ZSTD(1)),
